some progress on sync database, still not working
Database information are stored in SyncFile. Implemented a function to setup the SyncFile tree from database - this is done before each sync. Currently it is recursive, which is quite slow because it calls a sql query for each item.
Should think of a better way... Also the recursive algorithm doesnt check if the item is file or folder, sql is called anyway which is bad and waste of time.